+++ RegistroTecnicos/RegistroTecnicos/Models/Tecnicos.cs
@@ -16,7 +16,7 @@ public class Tecnicos
     [Required(ErrorMessage = "Este campo es obligatorio. ")]
     [Range(minimum: 0.1, maximum: 500000, ErrorMessage = "Por favor, ingrese una cantidad mayor a 0 y menor o igual a $500,000. ")]
 
-    public decimal SueldoHora { get; set; }
+    public double SueldoHora { get; set; }
 
     [ForeignKey("TiposTecnicosId")]
     public int TiposTecnicosId {  get; set; }
