+++ RegistroTecnicos/RegistroTecnicos/Models/TrabajoDetalle.cs
@@ -11,12 +11,12 @@ public class TrabajoDetalle
     [Required]
     public int Cantidad { get; set; }
 
-    public decimal Costo { get; set; }
+    public double Costo { get; set; }
 
     [Required(ErrorMessage = "Este campo es obligatorio. ")]
     [RegularExpression(@"^\d+(\.\d{1,2})?$", ErrorMessage = "Este campo solo permite números.")]
     [Range(minimum: 0.1, maximum: 500000, ErrorMessage = "Por favor, ingrese una cantidad mayor a 0 y menor o igual a $500,000. ")]
-    public decimal Precio { get; set; }
+    public double Precio { get; set; }
 
     //fk
     public int ArticuloId { get; set; }
